Added quick sort, "size_t" replaced with "long long"

+++ sort/selection/main.cpp
@@ -6,11 +6,11 @@ using namespace std;
 template<typename T>
 void selection_sort(vector<T>& vec) {
     //O(n^2)
-    size_t size = vec.size();
-    for (size_t i = 0; i < size; i++) {
-        size_t index = i;
+    long long size = vec.size();
+    for (long long i = 0; i < size; i++) {
+        long long index = i;
         T min = vec[index];
-        for (size_t j = i + 1; j < size; j++) {
+        for (long long j = i + 1; j < size; j++) {
             if (vec[j] < min) {
                 min = vec[j];
                 index = j;
